+++ packages/ui-transfer/src/components/web3Wallet/index.tsx
@@ -3,14 +3,13 @@ import { useAccount, useWalletConnector } from "@orderly.network/hooks";
 import { useTranslation } from "@orderly.network/i18n";
 import { ABSTRACT_CHAIN_ID_MAP } from "@orderly.network/types";
 import { Flex, Text, WalletIcon } from "@orderly.network/ui";
-import { formatAddress } from "../../utils";
 
 export const Web3Wallet: FC = () => {
   const { t } = useTranslation();
   const { wallet, connectedChain } = useWalletConnector();
   const { state: accountState, account } = useAccount();
 
-  const { walletName, address } = useMemo(() => {
+  const address = useMemo(() => {
     let address = accountState.address;
     if (
       connectedChain?.id &&
@@ -18,23 +17,21 @@ export const Web3Wallet: FC = () => {
     ) {
       address = account.getAdditionalInfo()?.AGWAddress;
     }
-    return {
-      walletName: wallet?.label,
-      address: formatAddress(address),
-    };
+
+    return address;
   }, [wallet, accountState, account, connectedChain]);
 
   return (
     <Flex justify="between">
       <Text size="sm" intensity={98}>
-        {t("transfer.web3Wallet")}
+        {t("transfer.web3Wallet.your")}
       </Text>
 
       <Flex gapX={1}>
-        <WalletIcon size={"xs"} name={walletName ?? ""} />
-        <Text size="sm" intensity={54}>
+        <WalletIcon size={"xs"} name={wallet?.label ?? ""} />
+        <Text.formatted size="sm" intensity={54} rule="address">
           {address}
-        </Text>
+        </Text.formatted>
       </Flex>
     </Flex>
   );