+++ packages/ui-transfer/src/utils.ts
@@ -1,9 +1,5 @@
+import { i18n } from "@orderly.network/i18n";
 import { API } from "@orderly.network/types";
-
-export function formatAddress(address?: string) {
-  if (!address) return "--";
-  return address.replace(/^(.{6})(.*)(.{4})$/, "$1......$3");
-}
 
 export const getTokenByTokenList = (tokens: API.TokenInfo[] = []) => {
   const tokenObj = tokens.reduce((acc, item) => {
@@ -21,3 +17,37 @@ export const feeDecimalsOffset = (origin?: number): number => {
 // export const priceDecimalsOffset = (origin?: number): number => {
 //   return Math.abs((origin ?? 2) - 5);
 // };
+
+export function checkIsAccountId(accountId: string) {
+  const regex = /^0x[a-fA-F0-9]{64}$/;
+  return regex.test(accountId);
+}
+
+export function getTransferErrorMessage(errorCode: number) {
+  if (errorCode === 34) {
+    return i18n.t("transfer.internalTransfer.error.transferInProgress");
+  }
+
+  if (errorCode === 17) {
+    return i18n.t("transfer.internalTransfer.error.withdrawalInProgress");
+  }
+
+  if (errorCode === 35) {
+    return i18n.t("transfer.internalTransfer.error.accountIdNotExist");
+  }
+
+  if (errorCode === 37) {
+    return i18n.t("transfer.internalTransfer.error.transferToSelf");
+  }
+
+  if (errorCode === 46) {
+    return i18n.t("transfer.internalTransfer.error.transferToSubAccount");
+  }
+
+  // TODO: when api return 500 status, not throw error code
+  // if (errorCode === -1000) {
+  //   return i18n.t("transfer.internalTransfer.error.scopeInsufficient");
+  // }
+
+  return i18n.t("transfer.internalTransfer.error.default");
+}