+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/i18n
 
+## 2.3.2
+
+### Patch Changes
+
+- 311e2a7: publish
+- publish
+
+## 2.3.2-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 2.3.1
+
+### Patch Changes
+
+- b2ca83a: publish
+- publish
+
+## 2.3.1-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 2.3.0
 
 ### Minor Changes