+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,31 @@
 # tsconfig
 
+## 0.6.2
+
+### Patch Changes
+
+- 311e2a7: publish
+- publish
+
+## 0.6.2-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 0.6.1
+
+### Patch Changes
+
+- b2ca83a: publish
+- publish
+
+## 0.6.1-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 0.6.0
 
 ### Minor Changes